backend/requirements.txt
flask
flask-cors
youtube-transcript-api
python-dotenv
google-generativeai>=0.8.5
langchain-google-genai>=0.0.10
langchain-community>=0.0.20
qdrant-client>=1.7.0
protobuf>=3.20.0,<5.0.0
deep-translator>=1.11.4
langdetect>=1.0.9